Use confidence intervals instead of standard deviation for uncertainties
This change decouples the uncertainties and the average in power spectrum plots.  In this way, the uncertainties are always plotted as the confidence interval centered on the median and the average is plotted independently using plt.scatter.  In extreme scenarios, it is possible for the average to lie outside a given confidence interval and, calculating errorbars for plotting in this scenario, results in a ValueError from negative values present in the errorbars.  Decoupling the two avoids this issue while accurately plotting both the average (expectation) and the confidence interval.

Other minor changes:
- Docstring fixes for "\D" in :math: to avoid SyntaxWarnings
- Replaced hasattr(x, "__iter__") with isinstance(x, Iterable)
- Other minor docstring fixes