Fix bug when nf and nt are None
If data_path points to a pyuvdata-compatible file, nf and nt are not required parameters.  They will default to the number of frequencies and times in the pyuvdata-compatible file, respectively.  However, nf is required if neta is None (default) to construct the sampler directory name.  This commit rearranges the order of operations in run_setup such that the visibilities are processed first and then the sampler directory is created.  The visibilities are now saved directly in run_setup instead of being saved in bayeseor.vis.preprocess_uvdata so the sampler directory can be created first (the default location for saving the visibility vector and/or instrument model if save_dir is None (default)) and then the visibilities/model can be saved to disk.